[POC] Remove the utils-legacy package
