+++ packages-next/keystone/package.json
@@ -33,7 +33,6 @@
     "@keystone-next/admin-ui-utils": "^5.0.2",
     "@keystone-next/fields": "^11.0.0",
     "@keystone-next/types": "^20.0.0",
-    "@keystone-next/utils-legacy": "^12.0.0",
     "@keystone-ui/button": "^5.0.0",
     "@keystone-ui/core": "^3.1.0",
     "@keystone-ui/fields": "^4.1.1",
