+++ packages-next/keystone/src/index.ts
@@ -2,3 +2,5 @@ export { createSystem } from './lib/createSystem';
 export { createExpressServer } from './lib/server/createExpressServer';
 export { initConfig } from './lib/config/initConfig';
 export { createApolloServerMicro } from './lib/server/createApolloServer';
+export { parseImageRef } from './lib/context/createImagesContext';
+export { parseFileRef } from './lib/context/createFilesContext';
