--- packages/utils/README.md
@@ -1,9 +1,0 @@
-<!--[meta]
-title: Utilities
-[meta]-->
-
-# Utilities
-
-This package is an internal Keystone package which contains various helper functions which are used throughout the monorepo.
-
-You should probably not use this directly in your Keystone projects.
