--- packages/utils/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@keystone-next/utils-legacy",
-  "description": "Common utility functions used throughout @keystonejs packages.",
-  "version": "12.0.0",
-  "author": "The KeystoneJS Development Team",
-  "license": "MIT",
-  "homepage": "https://github.com/keystonejs/keystone",
-  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/utils",
-  "main": "dist/utils-legacy.cjs.js",
-  "module": "dist/utils-legacy.esm.js",
-  "engines": {
-    "node": "^12.20 || >= 14.13"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.14.6",
-    "@keystone-next/types": "^20.0.0"
-  }
-}
