+++ tests/api-tests/package.json
@@ -27,7 +27,6 @@
   "dependencies": {
     "@keystone-next/test-utils-legacy": "^21.0.0",
     "@keystone-next/types": "^20.0.0",
-    "@keystone-next/utils-legacy": "^12.0.0",
     "apollo-cache-control": "^0.14.0",
     "express": "^4.17.1"
   }
